+++ R/conjugacy.R
@@ -32,7 +32,7 @@ conj_norm_mu <- function(y, tau, mu0 = 0, tau0 = 0.001, mult = 1) {
 conj_mvnorm_mu <- function(y, Q, mu0 = rep_len(0, n), Q0 = diag(0.001, n), mult = 1) {
   n <- nrow(y)
   newQ <- mult*Q0 + n*Q
-  newQ.inv <- my_inv(newQ)
+  newQ.inv <- chol_inv(newQ)
   MASS::mvrnorm(
     1,
     mu = newQ.inv %*% (Q0 %*% mu0 + Q %*% colSums(y)),
@@ -47,7 +47,7 @@ conj_matnorm_mu <- function(y, V, U = NULL, mu0, Q0) {
   if(is.null(U)) U <- diag(nrow(Q0) / nrow(V))
   VxU <- V %x% U
   newQ <- VxU + Q0
-  newQ.inv <- my_inv(newQ)
+  newQ.inv <- chol_inv(newQ)
   MASS::mvrnorm(
     1,
     mu = newQ.inv %*% (Q0 %*% mu0 + VxU %*% y),
@@ -59,7 +59,7 @@ conj_matnorm_mu <- function(y, V, U = NULL, mu0, Q0) {
 #' @export
 conj_lm_beta <- function(y, X, XtX = crossprod(X), tau, mu0, Q0) {
   newQ <- tau * XtX + Q0
-  newQ.inv <- my_inv(newQ)
+  newQ.inv <- chol_inv(newQ)
   MASS::mvrnorm(
     1,
     mu = newQ.inv %*% (Q0 %*% mu0 + tau*t(X) %*% y),
@@ -75,7 +75,7 @@ conj_matlm_beta <- function(y, X, V, U = NULL, mu0, Q0) {
 
   XtU <- if(is.null(U)) t(X) else crossprod(X, U)
   newQ <- V %x% (XtU %*% X) + Q0
-  newQ.inv <- my_inv(newQ)
+  newQ.inv <- chol_inv(newQ)
   MASS::mvrnorm(
     1,
     mu = newQ.inv %*% (Q0 %*% mu0 + (V %x% XtU) %*% y),
@@ -107,7 +107,7 @@ conj_norm_tau <- function(y, mu, a0 = 0.001, b0 = 0.001) {
 
 #' @rdname conjugacy
 #' @export
-conj_mvnorm_Q <- function(y, mu, V0, v0, V0_inv = my_inv(V0)) {
+conj_mvnorm_Q <- function(y, mu, V0, v0, V0_inv = chol_inv(V0)) {
   n <- nrow(y)
   stopifnot(identical(dim(y), dim(mu)))
   xmu <- y - mu
@@ -115,7 +115,7 @@ conj_mvnorm_Q <- function(y, mu, V0, v0, V0_inv = my_inv(V0)) {
     txmu <- xmu[i, , drop = FALSE]
     t(txmu) %*% txmu
   }))
-  V2 <- my_inv(V0_inv + tmp)
+  V2 <- chol_inv(V0_inv + tmp)
   rwish(
     V = V2,
     v = n + v0
@@ -124,14 +124,14 @@ conj_mvnorm_Q <- function(y, mu, V0, v0, V0_inv = my_inv(V0)) {
 
 #' @rdname conjugacy
 #' @export
-conj_matnorm_V <- function(y, mu, U = NULL, V0, v0, V0_inv = my_inv(V0)) {
+conj_matnorm_V <- function(y, mu, U = NULL, V0, v0, V0_inv = chol_inv(V0)) {
   if(!is.matrix(y) || !is.matrix(mu)) stop("'y' and 'mu' must be matrices")
   n <- nrow(mu)
   stopifnot(identical(dim(y), dim(mu)))
   ymu <- y - mu
   ymu2 <- if(is.null(U)) t(ymu) else crossprod(ymu, U)
 
-  V2 <- my_inv(V0_inv + ymu2 %*% ymu)
+  V2 <- chol_inv(V0_inv + ymu2 %*% ymu)
   rwish(
     V = V2,
     v = n + v0
@@ -146,7 +146,7 @@ conj_lm_tau <- function(y, X, beta, Xbeta = X %*% beta, a0 = 0.001, b0 = 0.001) 
 
 #' @rdname conjugacy
 #' @export
-conj_matlm_sigma <- function(y, X, beta, Xbeta = X %*% beta, U = NULL, V0, v0, V0_inv = my_inv(V0)) {
+conj_matlm_sigma <- function(y, X, beta, Xbeta = X %*% beta, U = NULL, V0, v0, V0_inv = chol_inv(V0)) {
   conj_matnorm_V(y = y, mu = Xbeta, U = U, v0 = v0, V0_inv = V0_inv)
 }