+++ NEWS.md
@@ -2,6 +2,8 @@
 
 - Export `logit()` and `expit()`
 
+- Rename `my_inv()` to `chol_inv()`.
+
 # v0.1.0
 
 Initial version of package